third_party/WebKit/LayoutTests/storage/indexeddb/idbcursor-continue-exception-order.html

<!DOCTYPE html>
<title>IndexedDB: IDBCursor continue() Exception Ordering</title>
<link rel="help" href="https://w3c.github.io/IndexedDB/#dom-idbcursor-continue">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="resources/testharness-helpers.js"></script>

<script>

indexeddb_test(
    function(t, db) {
        var s = db.createObjectStore('s');
        s.put('value', 'key');
    },
    function(t, db) {
        var s = db.transaction('s', 'readonly').objectStore('s');
        var r = s.openKeyCursor();
        r.onsuccess = t.step_func(function() {
            r.onsuccess = null;
            var cursor = r.result;
            setTimeout(t.step_func(function() {
                assert_throws('TransactionInactiveError', function() {
                    cursor.continue({not: "a valid key"});
                }, '"Transaction inactive" check (TransactionInactiveError) ' +
                   'should precede "invalid key" check (DataError)');
                t.done();
            }), 0);
        });
    },
    'IDBCursor.continue exception order: TransactionInactiveError vs. DataError'
);

indexeddb_test(
    function(t, db) {
        var s = db.createObjectStore('s');
        s.put('value', 'key');
    },
    function(t, db) {
        var s = db.transaction('s', 'readonly').objectStore('s');
        var r = s.openKeyCursor();
        r.onsuccess = t.step_func(function() {
            r.onsuccess = null;
            var cursor = r.result;
            cursor.continue();
            r.onsuccess = t.step_func(function() {
                setTimeout(t.step_func(function() {
                    assert_throws('TransactionInactiveError', function() {
                        cursor.continue();
                    }, '"Transaction inactive" check (TransactionInactiveError) ' +
                       'should precede "got value flag" check (InvalidStateError)');
                    t.done();
                }), 0);
            });
        });
    },
    'IDBCursor.continue exception order: TransactionInactiveError vs. InvalidStateError'
);

indexeddb_test(
    function(t, db) {
        var s = db.createObjectStore('s');
        s.put('value', 'key');
    },
    function(t, db) {
        var s = db.transaction('s', 'readonly').objectStore('s');
        var r = s.openKeyCursor();
        r.onsuccess = t.step_func(function() {
            r.onsuccess = null;
            var cursor = r.result;
            cursor.continue();
            assert_throws('InvalidStateError', function() {
                cursor.continue({not: "a valid key"});
            }, '"got value flag" check (InvalidStateError) should precede ' +
               '"invalid key" check (DataError)');
            t.done();
        });
    },
    'IDBCursor.continue exception order: InvalidStateError vs. DataError'
);

</script>
